Changes to data model and custom excel reader improvs

+++ RDS.ExpenseTracker.Data/Entities/EFinancialAccount.cs
@@ -5,8 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string? Description { get; set; }
-        public decimal Availability { get; set; }
-
+        public int Availability { get; set; }
         public ICollection<ETransaction>? Transactions { get; set; }
     }
 }
